alertlink: move inside core
