+++ static/app/views/monitors/components/monitorForm.tsx
@@ -3,8 +3,8 @@ import {css} from '@emotion/react';
 import styled from '@emotion/styled';
 import {Observer} from 'mobx-react';
 
-import AlertLink from 'sentry/components/alertLink';
 import {Alert} from 'sentry/components/core/alert';
+import AlertLink from 'sentry/components/core/alertLink';
 import {FieldWrapper} from 'sentry/components/forms/fieldGroup/fieldWrapper';
 import NumberField from 'sentry/components/forms/fields/numberField';
 import SelectField from 'sentry/components/forms/fields/selectField';
