+++ static/app/views/settings/projectAlerts/settings.tsx
@@ -1,7 +1,7 @@
 import {Fragment} from 'react';
 
-import AlertLink from 'sentry/components/alertLink';
 import {LinkButton} from 'sentry/components/button';
+import AlertLink from 'sentry/components/core/alertLink';
 import Form from 'sentry/components/forms/form';
 import JsonForm from 'sentry/components/forms/jsonForm';
 import LoadingError from 'sentry/components/loadingError';
